+++ src/test/java/ru/vyarus/guice/validator/ComposedCheckValidationTest.java
@@ -1,39 +1,42 @@
 package ru.vyarus.guice.validator;
 
-import com.google.inject.Guice;
-import com.google.inject.Injector;
-import org.junit.BeforeClass;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 import ru.vyarus.guice.validator.compositeannotation.ComposedCheckService;
 
 import javax.validation.ConstraintViolationException;
+import java.util.Collection;
 
 /**
  * @author Vyacheslav Rusakov
  * @since 25.06.2014
  */
-public class ComposedCheckValidationTest {
-    private static ComposedCheckService composedCheckService;
+@RunWith(Parameterized.class)
+public class ComposedCheckValidationTest extends AbstractParameterizedTest<ComposedCheckService> {
 
-    @BeforeClass
-    public static void setUp() throws Exception {
-        Injector injector = Guice.createInjector(new ValidationModule());
-        composedCheckService = injector.getInstance(ComposedCheckService.class);
+    public ComposedCheckValidationTest(String type, ComposedCheckService service) {
+        super(type, service);
+    }
+
+    @Parameterized.Parameters(name = "{index}: {0}")
+    public static Collection<Object[]> generateData() {
+        return AbstractParameterizedTest.generateData(ComposedCheckService.class);
     }
 
     @Test(expected = ConstraintViolationException.class)
     public void testValidationFail() throws Exception {
-        composedCheckService.checkParam("f");
+        service.checkParam("f");
     }
 
     @Test
     public void testValidation() {
-        composedCheckService.checkParam("valid string");
+        service.checkParam("valid string");
     }
 
     // todo can't make composed annotation work for return value
 //    @Test(expected = ConstraintViolationException.class)
 //    public void testReturn() {
-//        composedCheckService.checkReturn("valid string");
+//        service.checkReturn("valid string");
 //    }
 }