+++ src/test/java/ru/vyarus/guice/validator/CustomTypeValidationTest.java
@@ -4,10 +4,12 @@ import com.google.inject.Guice;
 import com.google.inject.Injector;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.junit.runners.Parameterized;
 import ru.vyarus.guice.validator.customtype.ComplexBean;
 import ru.vyarus.guice.validator.customtype.CustomService;
 
 import javax.validation.ConstraintViolationException;
+import java.util.Collection;
 
 /**
  * Check custom validator with injection.
@@ -15,22 +17,24 @@ import javax.validation.ConstraintViolationException;
  * @author Vyacheslav Rusakov
  * @since 24.06.2014
  */
-public class CustomTypeValidationTest {
-    private static CustomService customService;
+public class CustomTypeValidationTest extends AbstractParameterizedTest<CustomService> {
 
-    @BeforeClass
-    public static void setUp() throws Exception {
-        Injector injector = Guice.createInjector(new ValidationModule());
-        customService = injector.getInstance(CustomService.class);
+    public CustomTypeValidationTest(String type, CustomService service) {
+        super(type, service);
+    }
+
+    @Parameterized.Parameters(name = "{index}: {0}")
+    public static Collection<Object[]> generateData() {
+        return AbstractParameterizedTest.generateData(CustomService.class);
     }
 
     @Test(expected = ConstraintViolationException.class)
     public void testValidationFail() throws Exception {
-        customService.doAction(new ComplexBean("soso", 12));
+        service.doAction(new ComplexBean("soso", 12));
     }
 
     @Test
     public void testValidation() throws Exception {
-        customService.doAction(new ComplexBean("perfect", 12));
+        service.doAction(new ComplexBean("perfect", 12));
     }
 }