+++ src/test/java/ru/vyarus/guice/validator/ScriptedValidationTest.java
@@ -4,11 +4,14 @@ import com.google.inject.Guice;
 import com.google.inject.Injector;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.junit.runners.Parameterized;
+import ru.vyarus.guice.validator.compositeannotation.ComposedCheckService;
 import ru.vyarus.guice.validator.script.ScriptedBean;
 import ru.vyarus.guice.validator.script.ScriptedService;
 
 import javax.validation.ConstraintViolationException;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.Date;
 
@@ -16,34 +19,36 @@ import java.util.Date;
  * @author Vyacheslav Rusakov
  * @since 25.06.2014
  */
-public class ScriptedValidationTest {
-    private static ScriptedService scriptedService;
+public class ScriptedValidationTest extends AbstractParameterizedTest<ScriptedService> {
 
-    @BeforeClass
-    public static void setUp() throws Exception {
-        Injector injector = Guice.createInjector(new ValidationModule());
-        scriptedService = injector.getInstance(ScriptedService.class);
+    public ScriptedValidationTest(String type, ScriptedService service) {
+        super(type, service);
+    }
+
+    @Parameterized.Parameters(name = "{index}: {0}")
+    public static Collection<Object[]> generateData() {
+        return AbstractParameterizedTest.generateData(ScriptedService.class);
     }
 
     @Test(expected = ConstraintViolationException.class)
     public void testParamsFail() throws Exception {
-        scriptedService.paramsValid(Collections.<Integer>emptyList(), 1);
+        service.paramsValid(Collections.<Integer>emptyList(), 1);
     }
 
     @Test
     public void testParams() {
-        scriptedService.paramsValid(Arrays.asList(1, 2, 3), 3);
+        service.paramsValid(Arrays.asList(1, 2, 3), 3);
     }
 
     @Test(expected = ConstraintViolationException.class)
     public void testScriptedParamFail() throws Exception {
         Date now = new Date();
-        scriptedService.validBean(new ScriptedBean(now, now));
+        service.validBean(new ScriptedBean(now, now));
     }
 
     @Test
     public void testScriptedParam() {
         Date now = new Date();
-        scriptedService.validBean(new ScriptedBean(now, new Date(now.getTime() + 10000)));
+        service.validBean(new ScriptedBean(now, new Date(now.getTime() + 10000)));
     }
 }