+++ src/test/java/ru/vyarus/guice/validator/matcher/ExceptionalService.java
@@ -1,0 +1,15 @@
+package ru.vyarus.guice.validator.matcher;
+
+import javax.validation.constraints.NotNull;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 20.12.2014
+ */
+@SuppressValidation
+public class ExceptionalService {
+
+    public void doSmth(@NotNull String val){
+
+    }
+}