+++ src/test/java/ru/vyarus/guice/validator/matcher/MatchedService.java
@@ -1,0 +1,12 @@
+package ru.vyarus.guice.validator.matcher;
+
+import javax.validation.constraints.NotNull;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 20.12.2014
+ */
+public class MatchedService {
+    public void doSmth(@NotNull String val){
+    }
+}